requirement.txt
markdown==3.3.4
python-pptx==0.6.19
